pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phishlens-tr"
version = "1.0.0"
description = "PhishLens TR - Şüpheli mesaj metinleri için kural tabanlı risk analizi (eğitim amaçlı)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "PhishLens TR"}]
keywords = ["phishing", "security", "turkish", "sms", "email"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "rich>=13.0.0",
    "requests>=2.28.0",
    "fpdf2>=2.7.0",
    "python-dotenv>=1.0.0",
    "mkdocs-material>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
full = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "openai>=1.0.0",
    "dnspython>=2.4.0",
]

[project.scripts]
phishlens = "phishlens.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
phishlens = ["data/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
